+ Offscreen code writing increment 2   started mac osx (quartz) impl. + added filename_setext API to securely set a filename ext 'inplace' (using strlcat) + started to port (previously disabled) Offscreen code in fluid   for testing new offscreen services
git-svn-id: http://seriss.com/public/fltk/fltk/branches/fctrunk@5755 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/AnsiWidget.cxx
@@ -106,7 +106,7 @@ static Color colors[] = {
 
 /*! Standard constructor for a widget.
  */
-AnsiWidget::AnsiWidget(int x, int y, int w, int h, int defsize) : 
+AnsiWidget::AnsiWidget(int x, int y, int w, int h, float defsize) : 
   Widget(x, y, w, h, 0) {
   labelsize(defsize);
   init();
